底盘2.3.2
修复功率限制后启动抖动问题

+++ RESTART/USER/main.c
@@ -1,9 +1,9 @@
 #include "main.h"
 
-int testnum1 = 0;
-int testnum2 = 0;
-int testnum3 = 0;
-int testnum4 = 0;
+float testnum1 = 0;
+float testnum2 = 0;
+float testnum3 = 0;
+float testnum4 = 0;
 
 int main(void)
 {
